abstract logic into custom hooks:keyNav, storeCell:
