+++ src/types/types.ts
@@ -1,0 +1,11 @@
+export interface Cell {
+  value: string
+  styles: string[]
+}
+
+export type CellProp = keyof Cell
+
+export interface CellRef {
+  row: number
+  col: number
+}
